add value edit modal
