+++ frontend/common/src/intl/locales/en/system.json
@@ -21,7 +21,7 @@
   "label.create-user": "Create User",
   "label.drug-name": "Drug Name",
   "label.drugs": "Drugs",
-  "label.edit-configuration": "Edit Configuration",
+  "label.edit-option": "Edit Option",
   "label.edit-user": "Edit User",
   "label.error": "Error",
   "label.forms": "Forms",
